+++ BE/HanaZoom/src/main/java/com/hanazoom/domain/member/dto/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.hanazoom.dto.member;
+package com.hanazoom.domain.member.dto;
 
 import lombok.Getter;
