+++ BE/HanaZoom/src/main/java/com/hanazoom/domain/member/dto/SignupRequest.java
@@ -1,4 +1,4 @@
-package com.hanazoom.dto.member;
+package com.hanazoom.domain.member.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
