+++ BE/HanaZoom/src/main/java/com/hanazoom/domain/member/repository/MemberRepository.java
@@ -1,13 +1,13 @@
-package com.hanazoom.repository;
+package com.hanazoom.domain.member.repository;
 
-import com.hanazoom.domain.member.Member;
+import com.hanazoom.domain.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 import java.util.UUID;
 
 public interface MemberRepository extends JpaRepository<Member, UUID> {
-    boolean existsByEmail(String email);
-
     Optional<Member> findByEmail(String email);
+
+    boolean existsByEmail(String email);
 }
