+++ BE/HanaZoom/src/main/java/com/hanazoom/global/config/WebConfig.java
@@ -1,4 +1,4 @@
-package com.hanazoom.config;
+package com.hanazoom.global.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
