+++ BE/HanaZoom/src/main/java/com/hanazoom/global/controller/HealthController.java
@@ -1,4 +1,4 @@
-package com.hanazoom.controller;
+package com.hanazoom.global.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
